Renamed cash to cache... And removed util package
