+++ modules/telegram/src/main/java/io/starter/telegram/generator/replykeyboard/rows/InlineKeyboardRowGenerator.java
@@ -1,4 +1,4 @@
-package io.starter.telegram.utils.generator.replykeyboard.rows;
+package io.starter.telegram.generator.replykeyboard.rows;
 
 import java.util.ArrayList;
 import java.util.List;
